Reflect PR feedback and fix ThePreStub unwinding
This change modifies handling of the failure to unwind managed exception to assert
and EEPOLICY_HANDLE_FATAL_ERROR(COR_E_EXECUTIONENGINE) as @jkotas has suggested.
It also adds a missing SetUnwindHasStarted() to the restored exception flags
after the 2nd pass in the interleaved exception unwinding so that the next
first pass step creates a new tracker instead of reusing the existing one
that was used for the 2nd pass. While that was working fine in my tests too,
I prefer using a clean tracker so that only the part of the state that
we know we need is transferred.
I have also fixed the missing section in the LEAF_END invocations in the context2.S.
As an additional change, I have fixed a problem that prevented unwinding of
assembler functions that use the PROLOG_WITH_TRANSITION_BLOCK macro.
The macro was using CFI annotation for the xmm registers and the libunwind
doesn't support that.
Moreover, the PAL_VirtualUnwind return status was not being checked in the
`Thread::VirtualUnwindToFirstManagedCallFrame`, so the failure resulted in an
infinite loop in there.
Finally, the `debug/di/amd64/floatconversion.S` was including the unixasmmacros.inc
using a relative path, which was now failing after I've moved it to a different place.
Since the new path is in the include paths, the fix was to remove the relative
path.