Implement FlushProcessWriteBuffers in PAL for Linux
This change implements FlushProcessWriteBuffers in PAL. Since there is
no API on Unix that would have the same effect, we simulate it by
sending a realtime signal to all threads. The signal handler does nothing
but signal the thread that called the API that it has run. Running the
handler ensures flushing the processor write buffers.