Fix loading of NI images on SELinux
This change fixes a problem that prevents us to load crossgen-ed managed
assemblies on SELinux when running in confined mode.
The problem was that when we load these images, we also apply relocations
to their sections and so we temporarily switch section protection from
RX to RW and then back. And the switching back (RW -> RX) is something
that SELinux doesn't allow.
I have investigated various ways to solve the issue, but found only
one way that works and doesn't require creating a copy of the file into
shared memory or an anonymous file. The way is to load all sections as
writable and keep them that way.