Add support for local copies of 3rd party dependencies
This change sets the RPATH of all the binaries we build to point
to $ORIGIN/3rdpartylibs. That enables using local copies of
the 3rd party dependencies just by placing them to the 3rdpartylibs
subfolder of the folder where the binaries are located. This
is the first place where the linker will look and if the folder is
empty or even doesn't exist, everything would behave as before this
change.