Disable activation injection during SIGSEGV handler startup
The SIGSEGV handler startup code runs on an alternate stack to be able to detect
and report stack overflows. It switches to the regular stack of the faulting
thread after that check.
If activation injection signal handler is ran during the period of time when the
SIGSEGV handler is running on the alternate stack, it runs on the alternate stack
too. That causes a stack overflow due to the limited size of the alternate stack
and the process crashes.

The fix is to disable the activation signal for the period when the SIGSEGV handler
is running on the alternate stack. While that means that the activation won't get
delivered, it is actually ok, since it would be rejected by the g_safeActivationCheckFunction
function anyways due to the fact that the thread was not running in managed code.