Fix bug in ProcessCLRException for InlinedCallFrame handling
The pICFSetAsLimitFrame was being passed by value to the
ExceptionTracker::ProcessOSExceptionNotification, but the function is
supposed to set it so that it can be used after catch handler returns to
reset the InlinedCallFrame to inactive state. This is needed to make
sure the frame is set to inactive even in case when exception is thrown
out of the pinvoked code.