Fix GetCurrentProcessCpuCount on Unix
The current implementation doesn't work correctly when NUMA is enabled
and the current process is not limited to run on a single NUMA node.
It relies on the GetProcessAffinityMask that returns 0 mask in such
case and so we return the processor count of 64.

This change uses process affinity as the primary source of truth on
platforms that support getting it. On other platforms, it uses the total
number of processors available on the machine.
In both cases, the value is limited by the cgroups imposed CPU count
limit.