Fix return address hijack in x86 / x64 probing loop
There is a problem in interaction between return address hijacking and
stack probing loop on x64 / x86. When we try to hijack return address of
a thread when the thread is executing the probing loop, uwinding fails
to unwind correctly to the caller and we find a wrong stack location of
the return address. This leads to a failure to hijack the return address
and a possible corruption of a local variable when we unhijack the
thread, as the slot that was incorrectly identified as the return
address location is in the middle of the stack frame of the method that
was hijacked and we overwrite it with a garbage that was at that
location before the locals were initialized.